requirements.txt
torch
torchvision
transformers
matplotlib<3.7
scanpy==1.9.1
scikit-learn
scipy
numpy
pandas==1.1.5
einops
local_attention
wandb
accelerate
boltons
braceexpand
google-cloud-storage
datasets
transformers
torch